This test
os::canonicalize_ok_if_file_exists

works by writing a file at a path A, canonicalizing A and then checking that the canonicalized path A is equal to A.  This assumes that the path to A is canonical in the first place, which it might not be if the temp directory the test is running isn't canonical.

Proposed fix: instead of comparing paths, compare file contents.